% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sunspots_}
\alias{sunspots_}
\title{Monthly Sunspot Data, 1749-2012}
\source{
Based on \link[datasets]{sunspot.month}
}
\description{
Modified version of the base \link[datasets]{sunspots} dataset.
Includes columns for year, month and start_date of the month.
Used for examples.
}
\keyword{data}
