% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{USA.presidents}
\alias{USA.presidents}
\title{First 44 US Presidents}
\source{
Based on
 \url{http://thedatahub.org/api/data/ba0cdb03-c0f0-45ff-a21f-63fdf6ce1a89/}
}
\description{
List of presidents. Includes date of inauguration, party and home state.
}
\keyword{data}
